Extract batchSplitter to it's own package
It's about to be used by the listener.
